requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
loguru>=0.6.0
Pillow>=9.0.0
PyQt5>=5.15.0
configparser>=5.3.0